+++ app/Http/Requests/CustomerRequest.php
@@ -8,6 +8,8 @@ use Illuminate\Foundation\Http\FormRequest;
 
 /**
  * 顧客作成のRequetクラス
+ *
+ * @package App\Http\Requests
  */
 class CustomerRequest extends FormRequest
 {
@@ -36,9 +38,10 @@ class CustomerRequest extends FormRequest
             'first_kana' => 'required|max:50',
             'gender' => 'required',
             'birthday' => 'required|date',
+            'post_code' => 'required','regex:/^[0-9]{3}-[0-9]{4}$/',
             'pref_id' => 'required',
             'city_id' => 'required',
-            'address' => 'required|max:80','regex:/^[0-9]{3}-[0-9]{4}$/',
+            'address' => 'required|max:80',
             'buildding' => 'max:80',
             'tel' => 'required','regex:/^0\d{1,3}-\d{1,4}-\d{4}$/',
             'mobile' => 'required','regex:/^(070|080|090)-\d{4}-\d{4}$/',
