+++ resources/views/create.blade.php
@@ -182,48 +182,22 @@
 
 @section('javascript')
 
-function setCities(prefId,cityId) {
-    if (!prefId) {
-        $("#city_id").empty();
-        return;
-    }
-
-    var request = $.ajax({
-        type: "GET",
-        url: "{{ route('pref.select', ['pref_id' => $pref->id]) }}",
-        data: {
-            "pref_id" : prefId
-        }
+    $(function() {
+        setCities($("#pref_id").val());
+    });
+    $("#pref_id").change(function() {
+        setCities($("#pref_id").val());
     });
 
-    request.done(function(responseData) {
-        $("#city_id").empty();
-        responseData.forEach(function(item, index){
-            if (cityId !== undefined && cityId === item.id) {
-                $("#city_id").append($('<option>').text(item.name).attr('value', item.id).prop('selected', true));
-            } else {
-                $("#city_id").append($('<option>').text(item.name).attr('value', item.id));
+
+
+    $("#complete").click(function() {
+        completeConfirm(function(result){
+            if (result) {
+                $("form").submit();
             }
         });
     });
-}
-
-$(function() {
-    setCities($("#pref_id").val());
-});
-$("#pref_id").change(function() {
-    setCities($("#pref_id").val());
-});
-
-
-
-$("#complete").click(function() {
-    completeConfirm(function(result){
-        if (result) {
-            $("form").submit();
-        }
-    });
-});
 
 @include('layouts.submit', ['btn' => '登録'])
